Check the index value type when making a sparse tensor
