+++ cpp/src/arrow/tensor/converter.h
@@ -25,10 +25,6 @@ namespace arrow {
 namespace internal {
 
 struct SparseTensorConverterMixin {
-  static Status CheckSparseIndexMaximumValue(
-      const std::shared_ptr<DataType>& index_value_type,
-      const std::vector<int64_t>& shape);
-
   static bool IsNonZero(const uint8_t val) { return val != 0; }
 
   static void AssignIndex(uint8_t* indices, int64_t val, const int elsize);
